backend/requirements.txt
fastapi
uvicorn
openai
sqlalchemy
psycopg2-binary
python-docx
python-multipart
pydantic
python-dotenv
groq
